Add `end_date_str`, sibling to `start_date_str`. (#147)
Some of the telemetry I am analyzing with Jetstream can be quite
delayed. Therefore, I'd like to accept enrollment events that occur
during the observation period, and not just during the enrollment
period. I expected to be able to use something like `end_date_str`,
but it doesn't exist. Presumably this is because on days before the
`end_date`, that `end_date_str` might not make sense. But querying for
data with an end date in the future shouldn't matter; it just won't be
present (until it is).

Co-authored-by: Anna Scholtz <anna@scholtzan.net>